+++ braindecode/models/__init__.py
@@ -33,7 +33,8 @@ from .eegnex import EEGNeX
 from .tsinception import TSceptionV1
 from .eegtcnet import EEGTCNet
 from .syncnet import SyncNet
-from .ifnet import IFNet
+from .fbcnet import FBCNet
+from .ifnet import IFNetV2
 
 # Call this last in order to make sure the dataset list is populated with
 # the models imported in this file.