+++ docs/whats_new.rst
@@ -52,6 +52,9 @@ Enhancements
 - Ensuring consistency on the expose of the activation function (:gh:`637` by `Bruno Aristimunha`_ )
 - Adding :class:`braindecode.models.SyncNet` (:gh:`643` by `Bruno Aristimunha`_ )
 - Adding :class:`braindecode.models.IFNet` (:gh:`660` by `Bruno Aristimunha`_ )
+- Creating the FilterBank module for new models (:gh:`656` by `Bruno Aristimunha`_ )
+- Including PytorchAudio as dependency and remove copied code (:gh:`661` by `Bruno Aristimunha`_)
+- Adding :class:`braindecode.models.FBCNet` (:gh:`657` by `Bruno Aristimunha`_ )
 
 Bugs
 ~~~~